+++ packages/core/src/matterbridgeEndpoint.ts
@@ -566,12 +566,44 @@ export class MatterbridgeEndpoint extends Endpoint {
   /**
    * Sets the state of the provided cluster on a given endpoint.
    *
+   * @param {Behavior.Type} cluster - The cluster to set.
+   * @param {Record<string, boolean | number | bigint | string | object | undefined | null>} value - The state to set for the cluster.
+   * @param {AnsiLogger} [log] - (Optional) The logger to use for logging the set. Errors are logged to the endpoint logger.
+   * @returns {Promise<boolean>} - A promise that resolves to a boolean indicating whether the cluster was successfully set.
+   *
+   * @remarks Requires matterbridge version 3.6.0 or higher.
+   *
+   * @remarks The overload that takes a Behavior.Type is typed.
+   */
+  async setCluster<T extends Behavior.Type>(cluster: T, value: Behavior.StateOf<T>, log?: AnsiLogger): Promise<boolean>;
+  /**
+   * Sets the state of the provided cluster on a given endpoint.
+   *
+   * @param {ClusterType | ClusterId | string} cluster - The cluster to set.
+   * @param {Record<string, boolean | number | bigint | string | object | undefined | null>} value - The state to set for the cluster.
+   * @param {AnsiLogger} [log] - (Optional) The logger to use for logging the set. Errors are logged to the endpoint logger.
+   * @returns {Promise<boolean>} - A promise that resolves to a boolean indicating whether the cluster was successfully set.
+   *
+   * @remarks Requires matterbridge version 3.6.0 or higher.
+   *
+   * @remarks The overload that takes a Behavior.Type is typed.
+   */
+  async setCluster(
+    cluster: ClusterType | ClusterId | string,
+    value: Record<string, boolean | number | bigint | string | object | undefined | null>,
+    log?: AnsiLogger,
+  ): Promise<boolean>;
+  /**
+   * Sets the state of the provided cluster on a given endpoint.
+   *
    * @param {Behavior.Type | ClusterType | ClusterId | string} cluster - The cluster to set.
    * @param {Record<string, boolean | number | bigint | string | object | undefined | null>} value - The state to set for the cluster.
    * @param {AnsiLogger} [log] - (Optional) The logger to use for logging the set. Errors are logged to the endpoint logger.
    * @returns {Promise<boolean>} - A promise that resolves to a boolean indicating whether the cluster was successfully set.
    *
    * @remarks Requires matterbridge version 3.6.0 or higher.
+   *
+   * @remarks The overload that takes a Behavior.Type is typed.
    */
   async setCluster(
     cluster: Behavior.Type | ClusterType | ClusterId | string,
@@ -584,11 +616,37 @@ export class MatterbridgeEndpoint extends Endpoint {
   /**
    * Retrieves the state of the provided cluster from the given endpoint.
    *
+   * @param {Behavior.Type} cluster - The cluster to retrieve the state from.
+   * @param {AnsiLogger} [log] - (Optional) The logger to use for logging the retrieve. Errors are logged to the endpoint logger.
+   * @returns {Record<string, boolean | number | bigint | string | object | undefined | null> | undefined} The state of the cluster, or undefined if the cluster is not found.
+   *
+   * @remarks Requires matterbridge version 3.6.0 or higher.
+   *
+   * @remarks The overload that takes a Behavior.Type is typed.
+   */
+  getCluster<T extends Behavior.Type>(cluster: T, log?: AnsiLogger): Behavior.StateOf<T> | undefined;
+  /**
+   * Retrieves the state of the provided cluster from the given endpoint.
+   *
+   * @param {ClusterType | ClusterId | string} cluster - The cluster to retrieve the state from.
+   * @param {AnsiLogger} [log] - (Optional) The logger to use for logging the retrieve. Errors are logged to the endpoint logger.
+   * @returns {Record<string, boolean | number | bigint | string | object | undefined | null> | undefined} The state of the cluster, or undefined if the cluster is not found.
+   *
+   * @remarks Requires matterbridge version 3.6.0 or higher.
+   *
+   * @remarks The overload that takes a Behavior.Type is typed.
+   */
+  getCluster(cluster: ClusterType | ClusterId | string, log?: AnsiLogger): Record<string, boolean | number | bigint | string | object | undefined | null> | undefined;
+  /**
+   * Retrieves the state of the provided cluster from the given endpoint.
+   *
    * @param {Behavior.Type | ClusterType | ClusterId | string} cluster - The cluster to retrieve the state from.
    * @param {AnsiLogger} [log] - (Optional) The logger to use for logging the retrieve. Errors are logged to the endpoint logger.
    * @returns {Record<string, boolean | number | bigint | string | object | undefined | null> | undefined} The state of the cluster, or undefined if the cluster is not found.
    *
    * @remarks Requires matterbridge version 3.6.0 or higher.
+   *
+   * @remarks The overload that takes a Behavior.Type is typed.
    */
   getCluster(
     cluster: Behavior.Type | ClusterType | ClusterId | string,